+++ apps/web/app/future/signup/page.tsx
@@ -1,5 +1,9 @@
 import Signup from "@pages/signup";
+import { withAppDir } from "app/AppDirSSRHOC";
 import { _generateMetadata } from "app/_utils";
+import { WithLayout } from "app/layoutHOC";
+
+import { getServerSideProps } from "@server/lib/signup/getServerSideProps";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -7,4 +11,4 @@ export const generateMetadata = async () =>
     (t) => t("sign_up")
   );
 
-export default Signup;
+export default WithLayout({ Page: Signup, getLayout: null, getData: withAppDir(getServerSideProps) })<"P">;
